fix long tests
